progress on adding sb databases
