+++ src/MAGEMin.c
@@ -813,15 +813,63 @@ global_variable ReadCommandLineOptions(	global_variable 	 gv,
 	else{
 		printf(" WARNING: Unknown research group '%s' has been provided, setting default one 'tc'\n",gv.research_group);
 		strcpy(gv.research_group,"tc");
-	}	
-
-	// checks if the end-member dataset option arg is correct, otherwise sets to default
-	if 	(gv.EM_dataset 	== -1 || gv.EM_dataset 	== 62  || gv.EM_dataset	== 633  || gv.EM_dataset == 634){
 	}
-	else{
-		gv.EM_dataset = -1;
-		printf(" WARNING: Unknown dataset '%d' has been provided, setting default one\n",gv.EM_dataset);
-	}	
+
+
+	// Set default option if provided dataset is wrong
+	if 	( strcmp(gv.research_group, "tc") 	== 0 ){
+
+		// checks if the end-member dataset option arg is correct, otherwise sets to default
+		if 	(gv.EM_dataset 	== -1 || gv.EM_dataset 	== 62  || gv.EM_dataset	== 633  || gv.EM_dataset == 634 || gv.EM_dataset == 635 || gv.EM_dataset == 636){
+		}
+		else{
+			gv.EM_dataset = -1;
+			printf(" WARNING: Unknown dataset '%d' has been provided, setting default one\n",gv.EM_dataset);
+		}
+
+		/* set-up database acronym here*/
+		if 		(strcmp(gv.db, "mp") 	== 0){
+			gv.EM_database = 0;
+		}
+		else if (strcmp(gv.db, "mb") 	== 0){
+			gv.EM_database = 1;
+		}
+		else if (strcmp(gv.db, "ig") 	== 0){
+			gv.EM_database = 2;
+		}
+		else if (strcmp(gv.db, "igad") 	== 0){
+			gv.EM_database = 3;
+		}
+		else if (strcmp(gv.db, "um") 	== 0){
+			gv.EM_database = 4;
+		}
+		else if (strcmp(gv.db, "ume") 	== 0){
+			gv.EM_database = 5;
+		}
+		else {
+			printf(" No or wrong database acronym has been provided, using default (metapelite [mp])\n");
+			gv.EM_database = 0;
+		}
+	}
+	else if( strcmp(gv.research_group, "sb") == 0 ){
+		// checks if the end-member dataset option arg is correct, otherwise sets to default
+		if 	(gv.EM_dataset 	== -1 || gv.EM_dataset 	== 2011 ){
+		}
+		else{
+			gv.EM_dataset = -1;
+			printf(" WARNING: Unknown dataset '%d' has been provided, setting default one\n",gv.EM_dataset);
+		}
+
+		/* set-up database acronym here*/
+		if 		(strcmp(gv.db, "sb11") 	== 0){
+			gv.EM_database = 0;
+		}
+		else {
+			printf(" No or wrong database acronym has been provided, using default Stixrude & Lithgow-Bertelloni 2011([sb11])\n");
+			gv.EM_database = 0;
+		}
+
+	}
 
 	if (gv.verbose == 1){		
 		printf("--verbose     : verbose              = %i \n", 	 	   		gv.verbose			);
@@ -847,34 +895,6 @@ global_variable ReadCommandLineOptions(	global_variable 	 gv,
 		printf("--mbCpx       : mbCpx                = %i \n", 	 	   		gv.mbCpx			);
 
 		printf("--out_matlab  : out_matlab           = %i \n", 	 	   		gv.output_matlab	);
-	}
-
-
-
-
-
-	/* set-up database acronym here*/
-	if 		(strcmp(gv.db, "mp") 	== 0){
-		gv.EM_database = 0;
-	}
-	else if (strcmp(gv.db, "mb") 	== 0){
-		gv.EM_database = 1;
-	}
-	else if (strcmp(gv.db, "ig") 	== 0){
-		gv.EM_database = 2;
-	}
-	else if (strcmp(gv.db, "igad") 	== 0){
-		gv.EM_database = 3;
-	}
-	else if (strcmp(gv.db, "um") 	== 0){
-		gv.EM_database = 4;
-	}
-	else if (strcmp(gv.db, "ume") 	== 0){
-		gv.EM_database = 5;
-	}
-	else {
-		printf(" No or wrong database acronym has been provided, using default (Igneous [ig])\n");
-		gv.EM_database = 2;
 	}
 
 	return gv;
