+++ src/TC_database/SS_xeos_PC_igad.c
@@ -10159,7 +10159,8 @@ struct ss_pc igad_g_pc_xeos[2450] = {
 {{0.894100,0.894100,0.745100,0.149100,0.000100}},
 {{0.894100,0.894100,0.894100,0.000100,0.000100}}};
 
-struct ss_pc igad_ilm_pc_xeos[1032] = {
+struct ss_pc igad_ilm_pc_xeos[1033] = {
+{{0.994833,0.118827,0.873424,0.988995}},    
 {{0.124010,0.000010,-0.007990,-0.007990}},
 {{0.124010,0.000010,0.116010,0.116010}},
 {{0.124010,0.124010,-0.007990,-0.007990}},
