+++ src/TC_database/TC_init_database.c
@@ -131,7 +131,7 @@ igneous_igad_dataset igneous_igad_db = {
 	{"spn"	,"cpx"	,"g"	,"ilm"	,"liq"	,"ol"	,"opx"	,"fsp"	,"lct"	,"mel"	,"ness"	,"kals"		},
 	
 	{1		,1		,1		,1		,1 		,1 		,1 		,1 		,1		,1		,1		,1			}, // allow solvus?
-	{3317	,4125	,2450	,1032	,4910	,381	,3412	,861	,21		,270	,1209	,21			}, // # of pseudocompound
+	{3317	,4125	,2450	,1033	,4910	,381	,3412	,861	,21		,270	,1209	,21			}, // # of pseudocompound
 	{0.195	,0.249	,0.145	,0.124	,0.15	,0.098	,0.249	,0.0249	,0.049	,0.19	,0.149 	,0.049		}, // discretization step
 
 	6.0, 						/** max dG under which a phase is considered to be reintroduced  					*/
