+++ src/TC_database/tc_gss_init_function.h
@@ -8,14 +8,12 @@
  **   Contact      : nriel[at]uni-mainz.de, kaus[at]uni-mainz.de
  **
  ** ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ @*/
-#ifndef __GSS_INIT_FUNCTION_H_
-#define __GSS_INIT_FUNCTION_H_
+#ifndef __TC_GSS_INIT_FUNCTION_H_
+#define __TC_GSS_INIT_FUNCTION_H_
 
 #include "../MAGEMin.h"
-
-typedef SS_ref (*SS_init_type) (	SS_ref 				 SS_ref_db,
-									global_variable 	 gv				);
-									
+#include "../initialize.h"
+								
 void TC_SS_init_mp(	                SS_init_type 		*SS_init,
 									global_variable 	 gv				);				
 void TC_SS_init_mb(	                SS_init_type 		*SS_init,
@@ -30,17 +28,5 @@ void TC_SS_init_um_ext(	            SS_init_type 		*SS_init,
 									global_variable 	 gv				);
 void TC_SS_init(	        	    SS_init_type 		*SS_init,
 									global_variable 	 gv				);
-
-SS_ref G_SS_init_EM_function(		SS_init_type   *SS_init,
-									int				ph_id,
-									SS_ref 			SS_ref_db,
-									char 		   *name, 
-									global_variable gv				);
-
-csd_phase_set CP_INIT_function(		csd_phase_set 	cp, 
-									global_variable gv				);
-									
-stb_system SP_INIT_function(		stb_system 		sp,
-									global_variable gv				);
 
 #endif
